backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
sqlalchemy
psycopg2-binary